updated training file
Updated training file for the higher F1 Scoring. If there is any bugs with this file please let me know and I will fix it to the best of my ability. as well, I have also done a full project file dump, so there will be files that do not execute properly for a little while. until i get all the files sorted out and either working or removed :D