Delete filtered_occlusion_test.py
This file was a part of a whole different project lol. sorry about that everyone this is a whole different beast of a project setting on my computer that I didn't think was going to work. Again, my apologies for this file getting thrown in the mix. it's been rough trying to organize this project and the 14 other files so please bear with me